Fix ruff lint failures that have blocked backend-ci and worker-ci since day one
Discovered while wiring the new Flutter frontend up to a real backend:
neither backend-ci nor worker-ci had ever gone green, all the way back
to the first commit — both run `ruff check .` as a separate step from
the test suite, so the actual pytest suites (worker 37/37, API 21/21
non-DB-gated) were passing the whole time while lint silently failed
the job.

- worker/services/imaging.py: `int(round(dpi[0]))` was flagged
  RUF046 (round() already returns int) — genuinely redundant, removed.
- api/pyproject.toml: ignore B008 (FastAPI's `Depends(...)` as a
  parameter default is the framework's documented idiom, not a bug —
  ruff's own docs list this exact ignore for FastAPI projects) and
  UP042/UP046 (the codebase's `(str, Enum)` mixins are intentional for
  storage/serialization per app/db/enums.py's own docstring — StrEnum
  changes `str()` semantics, so that's a behavior change, not a style
  fix; `Generic[T]` is kept for broad Pydantic v2 tooling compat).
- Remaining 35 findings (import sorting, `Union[X, Y]` -> `X | Y`,
  `datetime.timezone.utc` -> `UTC`) were mechanical and auto-fixed by
  `ruff check --fix`.

Verified locally in fresh venvs: `ruff check .` passes clean in both
services, worker pytest 37/37, API pytest 21/21 (30 DB-gated tests
skip locally as before, run in CI against real Postgres/MinIO).

+++ services/worker/worker/services/imaging.py
@@ -37,7 +37,7 @@ def extract_metadata(data: bytes) -> dict[str, Any]:
     """
     with Image.open(BytesIO(data)) as img:
         dpi = img.info.get("dpi")
-        source_dpi = int(round(dpi[0])) if dpi else None
+        source_dpi = round(dpi[0]) if dpi else None
         has_alpha = img.mode in ("RGBA", "LA", "PA") or "transparency" in img.info
         return {
             "width": img.width,
